Add API step definitions and remove unnecessary code
