--- packages/api/test/steps.js
@@ -1,1 +1,0 @@
-module.exports=require('../index')
